refactor(linter): move duplicate methods in utils
